bumped patchlevel to b104 fixed a couple of stary index()'s fix a cast in viconf fixed broken tools/Makefile.in

+++ tools/viconf.c
@@ -180,7 +180,7 @@ LockedFile(char *filename)
     }
 
   fileptr = fdopen(fd,"w");
-  fprintf(fileptr,"%d\n",getpid());
+  fprintf(fileptr,"%d\n",(int) getpid());
   fclose(fileptr);
   return (0);
 } /* LockedFile() */
